\name{SharkPack-package}
\alias{SharkPack-package}
\alias{SharkPack}
\docType{package}
\title{
\packageTitle{SharkPack}
}
\description{
\packageDescription{SharkPack}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{SharkPack}
\packageIndices{SharkPack}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{SharkPack}

Maintainer: \packageMaintainer{SharkPack}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
